Put QC workflow in dedicated task
