+++ PWGHF/Tasks/CMakeLists.txt
@@ -68,3 +68,8 @@ o2physics_add_dpl_workflow(task-sel-optimisation
                     SOURCES taskSelOptimisation.cxx
                     PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore
                     COMPONENT_NAME Analysis)
+
+o2physics_add_dpl_workflow(task-track-index-skim-mc-match-qc
+                    SOURCES taskTrackIndexSkimMcMatchQC.cxx
+                    PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore
+                    COMPONENT_NAME Analysis)
